refactor: 文档重构和分区选择算法bug修复
🐛 Bug修复:
- 修复Topic.GetPartitionForKey中哈希值负数导致的数组越界问题
- 使用int64类型处理哈希值避免索引越界

📁 文档重构:
- 按阶段重新组织文档结构: doc/stage1/, doc/stage2/
- 创建doc/README.md作为文档索引
- 移动阶段1相关文档到stage1目录
- 创建阶段2网络通信相关文档

📚 新增深度解答文档:
- Consumer分区消费机制的架构设计问题
- Kafka Offset持久化的三层管理机制解析
- Seek操作在网络架构中的实现策略
- TCP服务器实现的疑问解答

💡 学习价值:
- 深入理解分布式系统的状态管理挑战
- 掌握客户端-服务器架构中的职责分离
- 理解Kafka offset管理的核心设计原理

+++ internal/common/topic.go
@@ -54,7 +54,7 @@ func (t *Topic) GetPartitionForKey(key []byte) *Partition {
 	count := t.GetPartitionCount()
 	var h maphash.Hash
 	h.WriteString(string(key))
-	return t.Partitions[int32(h.Sum64())%count]
+	return t.Partitions[int64(h.Sum64())%int64(count)]
 }
 
 func (t *Topic) GetPartitionCount() int32 {
